+++ typescript/packages/plugins/jupiter/package.json
@@ -14,6 +14,7 @@
     "dependencies": {
         "@jup-ag/api": "6.0.31",
         "@goat-sdk/core": "workspace:*",
+        "@goat-sdk/wallet-solana": "workspace:*",
         "zod": "catalog:",
         "@solana/web3.js": "catalog:"
     },
