+++ typescript/packages/plugins/jupiter/src/jupiter.service.ts
@@ -1,0 +1,71 @@
+import { Tool } from "@goat-sdk/core";
+import { SolanaWalletClient } from "@goat-sdk/wallet-solana";
+import { createJupiterApiClient } from "@jup-ag/api";
+import { ComputeBudgetProgram, Connection, PublicKey, TransactionInstruction } from "@solana/web3.js";
+import { GetQuoteParameters } from "./parameters";
+
+export class JupiterService {
+    private readonly jupiterApiClient: ReturnType<typeof createJupiterApiClient>;
+
+    constructor() {
+        this.jupiterApiClient = createJupiterApiClient();
+    }
+
+    @Tool({
+        description: "Gets a quote for a swap on the Jupiter DEX.",
+    })
+    async getQuote(parameters: GetQuoteParameters) {
+        try {
+            const response = this.jupiterApiClient.quoteGet(parameters);
+
+            console.log(JSON.stringify(response, null, 2));
+            return response;
+        } catch (error: unknown) {
+            if (error && typeof error === "object" && "response" in error) {
+                const response = error.response as Response;
+
+                const result = await response.json();
+                console.error(result);
+                throw new Error(`Failed to get quote: ${result.error}`);
+            }
+            throw error;
+        }
+    }
+
+    @Tool({
+        description: "Swaps tokens on the Jupiter DEX.",
+    })
+    async swapTokens(walletClient: SolanaWalletClient, parameters: GetQuoteParameters) {
+        const quoteResponse = await this.getQuote(parameters);
+        const { swapInstruction, addressLookupTableAddresses } = await this.jupiterApiClient.swapInstructionsPost({
+            swapRequest: {
+                userPublicKey: walletClient.getAddress(),
+                quoteResponse: quoteResponse,
+            },
+        });
+
+        const deserializedInstruction = new TransactionInstruction({
+            programId: new PublicKey(swapInstruction.programId),
+            keys: swapInstruction.accounts.map((key) => ({
+                pubkey: new PublicKey(key.pubkey),
+                isSigner: key.isSigner,
+                isWritable: key.isWritable,
+            })),
+            data: Buffer.from(swapInstruction.data, "base64"),
+        });
+
+        // TODO: Make this dynamic?
+        const computeBudgetIx = ComputeBudgetProgram.setComputeUnitLimit({
+            units: 400000, // Adjust this value as needed
+        });
+
+        const { hash } = await walletClient.sendTransaction({
+            instructions: [computeBudgetIx, deserializedInstruction],
+            addressLookupTableAddresses,
+        });
+
+        return {
+            hash,
+        };
+    }
+}
